Processes the complete_shakespeare works to stack of words
The processing of sonnets is ugly. Should probably attach
the plays processed to the sonnets processed instead of doing it
my ugly way on sonnets.
If you find any mistakes please let me know. It's rather hard to
sweep for errors that one hasn't thought of because of how hard
a list of words is to read.
With this larger sample size, we should have a better grammar
and better "Shakespeare's Voice," than the opposition.